+++ src/onegov.core/onegov/core/browser_session.py
@@ -2,7 +2,7 @@ from dogpile.cache.api import NO_VALUE
 
 
 class BrowserSession(object):
-    """ A session bound to a token (sessionid cookie). Used to store data
+    """ A session bound to a token (session_id cookie). Used to store data
     about a client on the server.
 
     Instances should be called ``browser_session`` to make sure we don't
@@ -22,10 +22,21 @@ class BrowserSession(object):
 
     """
 
-    def __init__(self, namespace, token, cache):
+    def __init__(self, namespace, token, cache, on_dirty=None):
         self._namespace = namespace
         self._token = token
         self._cache = cache
+        self._on_dirty = on_dirty or self._on_dirty
+        self._is_dirty = False
+
+    def _on_dirty(self, namespace, token):
+        """ Called when the browser session gets dirty. That is, the first time
+        a change is made to the cache.
+
+        Use ``on_dirty`` in the :meth:`__init__` method to override.
+
+        """
+        pass
 
     def mangle(self, name):
         """ Takes the given name and returns a key bound to the namespace and
@@ -38,6 +49,14 @@ class BrowserSession(object):
     def has(self, name):
         """ Returns true if the given name exists in the cache. """
         return self._cache.get(self.mangle(name)) is not NO_VALUE
+
+    def get(self, name, default=None):
+        value = self._cache.get(self.mangle(name))
+
+        if value is NO_VALUE:
+            return default
+        else:
+            return value
 
     def __getattr__(self, name):
         result = self._cache.get(self.mangle(name))
@@ -53,14 +72,24 @@ class BrowserSession(object):
         else:
             self._cache.set(self.mangle(name), value)
 
+            if not self._is_dirty:
+                self._on_dirty(self._namespace, self._token)
+
+            self._is_dirty = True
+
     def __delattr__(self, name):
         if name.startswith('_'):
             super(BrowserSession, self).__delattr__(name)
         else:
             self._cache.delete(self.mangle(name))
 
+            if not self._is_dirty:
+                self._on_dirty(self._namespace, self._token)
+
+            self._is_dirty = True
+
     # act like a dict
-    def __getitem__(self, name):
+    def __getitem__(self, name, *args):
         result = self._cache.get(self.mangle(name))
 
         if result is NO_VALUE: